# Conflicts:
#	lib/dataprovider/auth/auth.dart
#	lib/main.dart
#	lib/models/auth/auth.dart
#	lib/screens/auth/signin.dart
#	lib/screens/report/earning/earning.dart
#	lib/screens/report/earning/tabview/daily.dart
#	lib/screens/report/earning/tabview/weekly.dart
#	lib/screens/report/summary/summary.dart
#	lib/screens/report/summary/tabview/daily.dart
#	lib/screens/report/summary/tabview/weekly.dart
#	lib/screens/resetpassword/phonenumber.dart
#	lib/screens/resetpassword/reset_password.dart
#	lib/screens/settings/settings.dart
#	lib/screens/wallet.dart
#	lib/widgets/barchart/earning_barchart.dart
#	lib/widgets/onmap/offlinemode.dart
#	lib/widgets/onmap/onlinnemode.dart
#	pubspec.yaml

+++ lib/helper/constants.dart
@@ -34,7 +34,7 @@ late LatLng droppOffLocation;
 late String distance;
 late String duration;
 late String price;
-late int balance;
+late double balance;
 String? passengerProfilePictureUrl;
 // final player = AssetsAudioPlayer();
 bool isAccepted = false;
